chore: support view mode in tags
